# src/go-service/docs/swagger.yaml
basePath: /
definitions:
  app.ErrorResponse:
    properties:
      error:
        example: item not found
        type: string
    type: object
  app.Item:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Apple
        type: string
      price:
        example: 1.5
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: storhetmax@mail.ru
    name: Евланичев Максим Юрьевич
  description: REST API для управления каталогом товаров. Часть лабораторной работы
    №10 — сравнение FastAPI (Python) и Gin (Go).
  title: Items API (Go Gin)
  version: 1.0.0
paths:
  /items:
    get:
      description: Возвращает полный список товаров в каталоге
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.Item'
            type: array
      summary: Список всех товаров
      tags:
      - items
  /items/{id}:
    get:
      description: Возвращает товар по его числовому идентификатору
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.ErrorResponse'
      summary: Товар по ID
      tags:
      - items
  /ping:
    get:
      description: Возвращает pong — используется для health-check
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка работоспособности
      tags:
      - health
swagger: "2.0"